Add LLM.py
Introduce the following features:
1. `is_connected()` -> tries to create a connection with 8.8.8.8 to
   check if user is connected.
2. `ask_llm_prompted()` -> this is a function that asks the llm on
   sugarai a question with a custom prompt.

NOTE: This endpoint used in `ask_llm_prompted` is
`ai.sugarlabs.org/ask-llm-prompted` which is not part of sugarai as of
this commit. This works because the instance running sugarai is
temporarily running a fork of sugarai that has this updated endpoint.
Find the fork here: https://github.com/mebinthattil/sugar-ai

This new endpoint helps bypass the default prompt that sugar-ai
currently uses. This is useful when different acivitives want to have
different system prompts. It also gives customization options for model parameters such
as temperature, max_length, top_p, top_k and repitition_penality.