Review - Improve startup time
I've made kokoro instantiation run in the background because I noticed
this affects the time it takes for the activity to start, the start time
has moved from ~6secs to ~3secs, the caveat here though is the welcome
message when the activity opens uses espeak, at which point kokoro has loaded
and is now available to be used and I think that's fine.

The indentation for the kokoro voice list has also been improved to
show consistency and it does make the code more readable.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>