Review - Remove unneeded toolbarbutton
The persona button and kokoro button didn't need to be toolbarbuttons,
it just increased the number of clicks it'll take to get to the functionality
they provide.
The callbacks for both buttons have also been renamed.

I've also removed the try...except clause that checks for personas.json because
there's no need to check for it when it exists in the directory because the
directory listing shows it there, and it'll also be bundled with the activity so
the check isn't needed.

I've moved loading kokoro into a thread so as not to freeze the activity when
it opens, as the kokoro pipeline instantiation takes some time and this just makes
it easier to use the activity before the instantiation completes.

The only downside is using espeak until the kokoro pipeline is instantiated, and I
don't think that's much of a downside considering that's better than having the
activity freeze until it's instantiated, or waiting long for the activity to start.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>