Review - Handle downloading voice without an internet connection gracefully.
When you try to download a kokoro voice without an internet connection, an error is raised
which is expected - huggingface_hub.errors.LocalEntryNotFoundError -, but this error is in the
logs which isn't visible to the user at first glance and this also wasn't handled.
A message is displayed to the user letting them know they can't download a voice and why.

The info_label also lingered before this change, now it no longer lingers, some bloat has also
been removed.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>